Updates from public libcsp.

+++ src/csp_services.c
@@ -187,7 +187,7 @@ void csp_memfree(uint8_t node, uint32_t timeout) {
 	uint32_t memfree;
         int err = csp_get_memfree(node, timeout, &memfree);
 	if (err == CSP_ERR_NONE) {
-		printf("Free Memory at node %u is %"PRIu32" bytes\r\n", (unsigned int) node, memfree);
+		printf("Free Memory at node %u is %"PRIu32" bytes\r\n", node, memfree);
 	} else {
 		printf("Network error\r\n");
 	}
@@ -211,7 +211,7 @@ void csp_buf_free(uint8_t node, uint32_t timeout) {
 	uint32_t size;
 	int err = csp_get_buf_free(node, timeout, &size);
 	if (err == CSP_ERR_NONE) {
-		printf("Free buffers at node %u is %"PRIu32"\r\n", (unsigned int) node, size);
+		printf("Free buffers at node %u is %"PRIu32"\r\n", node, size);
 	} else {
 		printf("Network error\r\n");
 	}
@@ -234,7 +234,7 @@ void csp_uptime(uint8_t node, uint32_t timeout) {
 	uint32_t uptime;
 	int err = csp_get_uptime(node, timeout, &uptime);
 	if (err == CSP_ERR_NONE) {
-		printf("Uptime of node %u is %"PRIu32" s\r\n", (unsigned int) node, uptime);
+		printf("Uptime of node %u is %"PRIu32" s\r\n", node, uptime);
 	} else {
 		printf("Network error\r\n");
 	}
